driver function updated

+++ modules/main.py
@@ -1,3 +1,4 @@
+import socket
 import sys
 
 from GDrive_storage_class import GDrive_Storage
@@ -33,54 +34,64 @@ class CloudStorageManager:
         self.dropbox_account.unlink_account()
         self.dropbox_account_status = 0
 
-    def _get_filename_from_backup_path(self, path):
-        """Function for internal use only, to get a filename from the local path to it."""
-        name = []
-        counter = 0
-        for i in path[::-1]:
-            if counter and i != '/':
-                name.append(i)
-            if i == '/':
-                counter += 1
-            if counter > 1: break
-        return "".join(name[::-1])
+    def _connected(self):
+        try:
+            socketInstance = socket.create_connection(("www.google.com", 80))
+            socketInstance.close()
+            return True
+        except:
+            return False
+
+    def _alert(self):
+        if not (self.google_drive_account_status or self.dropbox_account_status):
+            print("None of the storage account was connected.")
+            return True
+        else:
+            return False
 
     def upload_file(self, local_path, backup_path, storage):
         """Upload file to one of the user's cloud storages"""
-        if storage:
+        if storage and self.google_drive_account_status:
             if backup_path:
-                backup_path = [i for i in backup_path.split('/')[1:] if i]
+                backup_path = [i for i in backup_path.split('/') if i]
                 self.google_drive_account.upload_file(local_path, backup_path)
             else:
                 self.google_drive_account.upload_file(local_path)
         else:
-            backup_path += '/' + local_path.split('/')[-1]
-            self.dropbox_account.upload_file(local_path, backup_path)
+            if self.dropbox_account_status:
+                backup_path += '/' + local_path.split('/')[-1]
+                self.dropbox_account.upload_file(local_path, backup_path)
 
     def download_file(self, filename, storage):
         """Download file from one of the user's cloud storages"""
-        if storage == 1:
+        if storage == 1 and self.google_drive_account_status:
             self.google_drive_account.download_file(filename)
-        elif storage == 0:
+        elif storage == 0 and self.dropbox_account_status:
             self.dropbox_account.download_file(filename)
         else:
-            self.google_drive_account.download_file(filename)
-            self.dropbox_account.download_file(filename)
+            if self.google_drive_account_status:
+                self.google_drive_account.download_file(filename)
+            if self.dropbox_account_status:
+                self.dropbox_account.download_file(filename)
 
     def sync(self, filename, storage):
         """Update file's content in user's cloud storages"""
-        if storage == 1:
+        if storage == 1 and self.google_drive_account_status:
             self.google_drive_account.sync(filename)
-        elif storage == 0:
+        elif storage == 0 and self.dropbox_account_status:
             self.dropbox_account.sync(filename)
         else:
-            self.google_drive_account.sync(filename)
-            self.dropbox_account.sync(filename)
+            if self.google_drive_account_status:
+                self.google_drive_account.sync(filename)
+            if self.dropbox_account_status:
+                self.dropbox_account.sync(filename)
 
     def delete_file(self, filename):
         """Delete file in user's cloud storages"""
-        self.dropbox_account.delete_file(filename)
-        self.google_drive_account.delete_file(filename)
+        if self.dropbox_account_status:
+            self.dropbox_account.delete_file(filename)
+        if self.google_drive_account_status:
+            self.google_drive_account.delete_file(filename)
 
     def list_files(self):
         """Write to the std all files in user's cloud storages"""
@@ -103,11 +114,9 @@ class CloudStorageManager:
             7: self.disconnect_dropbox_storage,
             8: self.disconnect_google_drive
         }
-
         print(
             """
-                Menu
-                
+                Menu\n\n
     1: Connect Dropbox account.\n
     2: Connect GoogleDrive account.\n
     3: Upload File.\n
@@ -117,19 +126,33 @@ class CloudStorageManager:
     7: Disconnect Dropbox account.\n
     8: Disconnect Google Drive account.\n
     9: List files.\n
-    10: Exit
-"""
-        )
-        option = int(input("Enter: "))
+    10: Exit""")
+        try:
+            option = int(input("Enter: "))
+        except:
+            return 0
         if not (1 <= option <= 10):
             print('Invalid action chosen')
-            return 0
         elif option == 10:
+            if self.google_drive_account_status:
+                self.disconnect_google_drive()
+            if self.dropbox_account_status:
+                self.disconnect_dropbox_storage()
             sys.exit(0)
-        elif option in [1, 2, 7, 8]:
+        elif option in (1, 2):
             option_funct[option]()
+        elif option == 7:
+            if not self.dropbox_account_status:
+                return print("Dropbox storage wasn't connected.")
+            option_funct[7]()
+        elif option == 8:
+            if not self.google_drive_account_status:
+                return print("Google Drive storage wasn't connected.")
+            option_funct[8]()
+        elif self._alert():
+            return
         elif option == 6:
-            self.delete_file(input('Enter the name of file to delete: '))
+            return self.delete_file(input('Enter the name of file to delete: '))
         else:
             if option == 3:
                 storage = int(input("Google Drive - 1, Dropbox - 0 (skip if unimportant)? "))
@@ -150,7 +173,10 @@ class CloudStorageManager:
     def run(self):
         """Run the main routine"""
         while True:
-            self.menu()
+            if self._connected():
+                self.menu()
+            else:
+                return print("No internet connection.")
 
 
 if __name__ == "__main__":
